+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -24,6 +24,7 @@ CHANGELOG
  * Add the `session:clear` command to remove all sessions from the configured handler
  * Generate JSON schema for YAML configuration of bundles
  * Add support for rate limited transports
+ * Register the security functions `is_granted()`, `is_authenticated()`, `is_fully_authenticated()` and `is_remember_me()` in the validator expression language
 
 8.1
 ---